Renamed Leetcode folder , created leetcodejs folder
